Testing out memory changes

+++ src/CommunityMatrix.cpp
@@ -12,27 +12,27 @@ void CommunityMatrix::InitializeMatrix() {
 
     row = communityMatrix.nrow();
     col = communityMatrix.ncol();
-    rowAbundance = std::vector<std::vector<int64_t>>(row);
-    eligibleRowIndexes = std::vector<std::vector<int64_t>>(row);
-    allIndexes = std::vector<std::vector<int64_t>>(row);
-    sums = std::vector<int64_t>(row);
+    rowAbundance = std::vector<std::vector<uint>>(row);
+    eligibleRowIndexes = std::vector<std::vector<uint>>(row);
+    allIndexes = std::vector<std::vector<uint>>(row);
+    sums = std::vector<uint>(row);
     Rcpp::NumericMatrix resultantMatrix(row, col);
     for(int i = 0; i < row; i++) {
         Rcpp::NumericVector community = communityMatrix(i, Rcpp::_);
-        std::vector<int64_t> communityVector = Rcpp::as<std::vector<int64_t>>(community);
+        std::vector<uint> communityVector = Rcpp::as<std::vector<uint>>(community);
         const size_t size = communityVector.size();
         eligibleRowIndexes[i].reserve(size);
         rowAbundance[i].reserve(size);
         sums[i] = std::accumulate(communityVector.begin(), communityVector.end(), 0LL);
-        allIndexes[i] = std::vector<int64_t>(sums[i]);
+        allIndexes[i] = std::vector<uint>(sums[i]);
         size_t index = 0;
-        int64_t currentPosition = 0;
+        uint currentPosition = 0;
         for(size_t j = 0; j < size; j++) {
-            int64_t val = communityVector[j];
+            uint val = communityVector[j];
             if(val > 0) {
                 eligibleRowIndexes[i].emplace_back(j);
                 rowAbundance[i].emplace_back(val);
-                for (int64_t k = 0; k < val; k++) {
+                for (uint k = 0; k < val; k++) {
                     allIndexes[i][index++] = currentPosition;
                 }
                 currentPosition++;
@@ -41,8 +41,8 @@ void CommunityMatrix::InitializeMatrix() {
     }
 }
 
-std::vector<int64_t> CommunityMatrix::GetCommunityMatrixByRow(const int row) const {
+std::vector<uint> CommunityMatrix::GetCommunityMatrixByRow(const int row) const {
     Rcpp::NumericVector community = communityMatrix(row, Rcpp::_);
-    const std::vector<int64_t> communityVector = Rcpp::as<std::vector<int64_t>>(community);
+    const std::vector<uint> communityVector = Rcpp::as<std::vector<uint>>(community);
     return communityVector;
 }
